Mover todo a funciones en modulo
